fix: \lz4\programs\lz4io.c(2873,13): warning C4777: 'fprintf' : format string '%llu' requires an argument of type 'unsigned __int64', but variadic argument 2 has type 'size_t' \lz4\programs\lz4io.c(2873,13): message : the sizes of types 'size_t' and 'unsigned __int64' might differ on other platforms \lz4\programs\lz4io.c(2873,13): message : consider using '%zu' in the format string \lz4\programs\lz4io.c(2873,13): warning C4777: 'fprintf' : format string '%llu' requires an argument of type 'unsigned __int64', but variadic argument 3 has type 'size_t' \lz4\programs\lz4io.c(2873,13): message : the sizes of types 'size_t' and 'unsigned __int64' might differ on other platforms \lz4\programs\lz4io.c(2873,13): message : consider using '%zu' in the format string

+++ programs/lz4io.c
@@ -2870,7 +2870,7 @@ int LZ4IO_displayCompressedFilesInfo(const char** inFileNames, size_t ifnIdx)
         }
         if (g_displayLevel >= 3) {
             /* verbose mode */
-            DISPLAYOUT("%s(%llu/%llu)\n", cfinfo.fileName, (unsigned long long)idx + 1, (unsigned  long long)ifnIdx);
+            DISPLAYOUT("%s(%zu/%zu)\n", cfinfo.fileName, idx + 1, ifnIdx);
             DISPLAYOUT("    %6s %14s %5s %8s %20s %20s %9s\n",
                         "Frame", "Type", "Block", "Checksum", "Compressed", "Uncompressed", "Ratio");
         }
